fix: amend test to work with different response type
