+++ app/routers/word_of_day.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends
 
 from app.crud.unit_of_work import UnitOfWork, get_unit_of_work
 from app.models.user import User
-from app.schemas.word_of_day import WordOfDayResponse
+from app.schemas.word import WordPublicWithPhonemes
 from app.services.word import WordService
 from app.users import current_active_user
 
@@ -10,11 +10,9 @@ from app.users import current_active_user
 router = APIRouter()
 
 
-@router.get("/api/v1/word_of_day", response_model=WordOfDayResponse)
+@router.get("/api/v1/word_of_day", response_model=WordPublicWithPhonemes)
 async def get_word_of_day(
     uow: UnitOfWork = Depends(get_unit_of_work), user: User = Depends(current_active_user)
-) -> WordOfDayResponse:
+) -> WordPublicWithPhonemes:
     today_word = uow.word_of_day.get_word_of_day()
-    return WordOfDayResponse(
-        word=WordService(uow).to_public_with_phonemes(today_word.word),
-    )
+    return WordService(uow).to_public_with_phonemes(today_word.word)
