--- app/schemas/word_of_day.py
@@ -1,7 +1,0 @@
-from pydantic import BaseModel
-
-from app.schemas.word import WordPublicWithPhonemes
-
-
-class WordOfDayResponse(BaseModel):
-    word: WordPublicWithPhonemes
